Update venv pickle files fixing tox tests
The pickle files for test virtualenv data were created using an older
version of pip. As tox started using the newer version of pip, these
files were obsolete and it lead to failure of tests run via tox. This is
now fixed by regenerating the pickle files for pip ver. 6.0.8