Rename 'test_pipdeptree.py' to 'test_integration.py'
So that a new 'test_pipdeptree.py' can be added for unit tests (with
mocking etc.).

Integration test suite will still be run during 'make test' but the
state of the test suite is pretty bad at present. In future, we might
replace it with a complete approach for easy maintenance.