Check for conflicts in 'show_only' and 'exclude' after handling case
Earlier, in the Tree.filter method, we were checking for conflicts in
'show_only' and 'exclude' sets before normalizing all values to lower
case as a result of which it would miss values that were in both the
sets but in different cases.

This is now fixed by checking for conflicts after normalizing all the
values in both sets to lowercase first.