Improve graphviz output to differentiate missing dependencies
The missing dependencies will now be displayed with 'dashed' nodes and
edges.

This commit also fixes the problem of duplicate nodes being displayed
in the graph, which would happen when the 'project_name' attribute of
'ReqPackage' and the corresponding 'DistPackage' obj different. Now we
are using the 'key' attribute which is guaranteed to be the same for
both objects.