Refactor PackageDAG code for reusing the node index/mapping
Earlier the rendering functions eg. 'render_text', 'render_json_tree'
were constructing their own node index/mapping. Now they are using the
one already existing in the tree instance.

To facilitate this, the 'PackageDAG' class has been extended with a
'sort' method that returns a new DAG of the same type but with it's
'_obj' dict sorted alphabetically by the node key. This the tree can
be sorted without converting it into a plain OrderedDict.