Deduplicate ReqPackage objects for extra requires
This handles the case where the same package is part of 2 different
extra_require profiles for eg.

  {'rsa': ['cryptography'],
   'signals': ['blinker'],
   'signedtoken': ['cryptography', 'pyjwt>=1.0.0']}

The above is the extra_require dict for oauthlib. Since 'cryptography'
is an extra require in multiple profiles, this caused duplicate
ReqPackage objects to be present in the tree.

As a result of deduplicating, no additional filtering is required in
presentation logic (ie. 'render_text' and 'dump_graphviz' functions).