Ensure packages specified using the '-p' option are always shown
If a package is specified using the '-p' option and it also happens to
be one of the dependencies of another package (also specified as part
of '-p' option), then the top level package itself is skipped from the
output.

For eg. consider following call

  $ pipdeptree -p 'abc,tox'

Now if 'tox' also happens to be the dependency of 'abc', then it's not
shown in the output unless '-a' flag (show all) is also specified.

This seems to be an old bug that has surfaced now with support for
extra requirements. There are cases where a package can be a dependent
of itself via extra requirements eg. tox depends on pluggy and
pluggy[dev] also depends on tox.