models/package: move `requires()` and `version` to `DistPackage`
This change exists due to how we use ReqPackage. Since we pass a
pip._vendor.pkg_resources.Requirement to Package's constructor
instead of the expected DistInfoDistribution, using this method and
property will cause a crash because Requirement does not have this
data.

To resolve this, they have been moved to DistPackage as it makes more
sense for the behavior to exist here rather than force ReqPackage to
also have this behavior.