Split freeze and text output into their own implementations (#459)
This snapshot also adds tests for the freeze output (since they did not
exist before). This refactor is done because of the following:
- The freeze output is not bound to change
- The text output is bound to change. This has been evident with
features like unicode support, license metadata, and future support for
more metadata and listing the deps of explicit extras. Having to ensure
that freeze is still working as expected after such changes makes it
harder to maintain

Though there is some code that is similar to the text output
implementation, splitting this up will ensure that changes to the text
output implementation won't cause problems to the freeze output
implementation.