Replace `pytest-console-scripts` plugin with smaller in-house solution (#476)
Resolves #474.

Due to us using the stdlib `importlib.metadata` and
`pytest-console-scripts` using the backport version for Python versions
< 3.10, this has caused some problems when `importlib_metadata` upgrades
to a version that makes changes which is incompatible with the stdlib
version.

After analyzing this plugin and the stdlib version of
`importlib.metadata`, I ultimately decided to remove this plugin in
favor of using `Distribution.entry_points()` to get the console script.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>