Implement --exclude-dependencies (#488)
Resolves #453.

Much harder than I expected to implement, but I think my logic and
implementation checks out. The algorithm is the following:

- Take the _exclude_ set from CLI input and resolve it to actual nodes
in the graph
- Determine all possible candidates for exclusion by performing DFS
- So we add the direct and transitive dependencies of the exclude nodes
we gathered in Step 1 into a `candidates` set, but **we aren't sure
yet** if they should be excluded since a non-excluded node maybe using
it. **This is the reason** why we have more steps below
- Create a reversed graph so that we know all the nodes that depend on a
specific node
- Iterate over `candidates`, **counting** the number of dependents for
each node using the reversed graph
- Now iterate over the explicitly-excluded nodes that we created in Step
1, where for each node:
  - Add it to the **new excluded set**
- Iterate over their children. For each child, get the dependent count
of that child and decrement it by 1
- If the count equals 0, we haven't added it to the new set, and it
isn't apart of the original resolved set created in Step 1, this child
is a candidate for exclusion so let's add it (and in effect, also add
their applicable children)
- If this specific branch is never reached, it means that **their is a
non-excluded node that has an outgoing edge to child**. We should not
remove it
- Return the **new excluded set**, which will contain both the
explicitly-excluded nodes and the dependency nodes associated to them