Allow --exclude to be used with --include (#491)
#273 removed the ability to use --include and --exclude together. I now
disagree with this change since not only was that a (accidental)
breaking change but I now see it makes sense to allow them to be used
together (e.g using pattern matching: `pidpeptree -p pytest* -e
pytest-mock`).

This change allows them to be used together and adds a few tests to
ensure that they continue to do so.